32 - 64

+++ Class_1CD.cpp
@@ -392,7 +392,7 @@ void memblock::add_block()
 
 //---------------------------------------------------------------------------
 //uint32_t memblock::get_numblocks()
-uint32_t memblock::get_numblocks()
+uint64_t memblock::get_numblocks()
 {
 	return numblocks;
 }
